requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
ftfy>=6.1.0
sentencepiece>=0.1.99
Pillow>=10.0.0
tqdm>=4.66.0
opencv-python>=4.8.0
imagehash>=4.3.1
numpy>=1.24.0
scipy>=1.11.0

# Vector search
faiss-cpu>=1.7.4
hnswlib>=0.7.0

# Web API (Day 2: Batch Recognition Service)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Data fetching
aiohttp>=3.8.0
aiofiles>=23.0.0
requests>=2.31.0
ijson>=3.2.0

# Database
sqlalchemy>=2.0.0

# CLI
click>=8.1.7

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

